+++ src/Main/Main.cpp
@@ -182,7 +182,7 @@ int                  CHE_GPU_NPGROUP;
 #ifdef GRACKLE_H2_SOBOLEV
 double              *H2_Op_T_Table     = NULL ;
 double              *H2_Op_Alpha_Table = NULL ; 
-double               H2_Op_N_elem      = 500  ;
+int                  H2_Op_N_elem      = 500  ;
 #endif
 
 #endif // SUPPORT_GRACKLE
